Revert code of useAppData hook back into App.js, to avoid either a constant re-render, or an inability to re-render properly with updateCommentsList. TODO: Find more elegant solution
